fix(practice): refactoring widgets to game/widgets
