+++ analysis_options.yaml
@@ -1,9 +1,1 @@
 include: package:strict_dart/analysis_options.yaml
-linter:
-  rules:
-    prefer_const_constructors: false
-    prefer_const_constructors_in_immutables: false
-    prefer_const_literals_to_create_immutables: false
-    prefer_single_quotes: false
-    prefer_final_fields: false
-    avoid_classes_with_only_static_members: false
