+++ lib/module/game/practice_mode/practice_mode_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-import '../../presentation/dimensions.dart';
-import '../../presentation/game/tap_and_key_listener.dart';
-import '../../presentation/game/traffic_light.dart';
+import '../../../presentation/dimensions.dart';
+import '../widgets/tap_and_key_listener.dart';
+import '../widgets/traffic_light.dart';
 
 class PracticeModePage extends StatefulWidget {
   static const route = 'practice_mode_page';
@@ -18,7 +18,7 @@ class _PracticeModePageState extends State<PracticeModePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Practice mode'),
+        title: const Text('Practice mode'),
       ),
       body: Center(
         child: Container(
@@ -26,7 +26,7 @@ class _PracticeModePageState extends State<PracticeModePage> {
           padding: AppDimensions.allPagePadding,
           child: Column(
             mainAxisAlignment: AppDimensions.containerMainAxisAlignment,
-            children: [
+            children: const [
               TrafficLight(),
               SizedBox(height: 40),
               TapAndKeyListener(),
