+++ lib/module/game/widgets/custom_keyboard_listener.dart
@@ -15,20 +15,6 @@ class CustomKeyboardListener extends StatefulWidget {
 
 class _CustomKeyboardListenerState extends State<CustomKeyboardListener> {
   FocusNode focusNode = FocusNode();
-  bool _hasPressedSpaceDown = false;
-  bool _hasPressedOtherKeyDown = false;
-
-  void _toggleHasPressedSpaceDown() {
-    setState(() {
-      _hasPressedSpaceDown = !_hasPressedSpaceDown;
-    });
-  }
-
-  void _toggleHasPressedOtherKeyDown() {
-    setState(() {
-      _hasPressedOtherKeyDown = !_hasPressedOtherKeyDown;
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -40,11 +26,6 @@ class _CustomKeyboardListenerState extends State<CustomKeyboardListener> {
             event.data.logicalKey == LogicalKeyboardKey.space;
         if (hasPressedSpace && event is RawKeyUpEvent) {
           widget.onAction();
-          _toggleHasPressedSpaceDown();
-        } else if (hasPressedSpace && event is RawKeyDownEvent) {
-          _toggleHasPressedSpaceDown();
-        } else {
-          _toggleHasPressedOtherKeyDown();
         }
       },
       child: widget.child,
